Cache PNGs under cache directory
Since metro caches the asset data, the plugin also needs to cache
the generated images. In order to decrease clutter, we also
move the generated images under a separate cache directory.

Currently, the cache directory is never cleaned, and only grows
over time. A solution for this will be proposed later.